# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Mapp3D Project CRUD",
    "description": "Mapp3D Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://ec2-3-25-145-248.ap-southeast-2.compute.amazonaws.com:3000"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/objects/": {
      "get": {
        "tags": [
          "Objects"
        ],
        "summary": "Get a list of objects",
        "description": "Get a list of objects in MongoDB",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Retreived all objects"
                    },
                    "count": {
                      "type": "integer",
                      "example": 1
                    },
                    "data": {
                      "type": "array",
                      "example": [
                        {
                          "_id": "61e4f519d450770ffab60d9c",
                          "objectName": "Object",
                          "objectURL": "https://sample-url.com",
                          "scale": "50 50 50",
                          "rotation": "0 0 0",
                          "position": "50 0 50",
                          "__v": 0
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error message"
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/objects/{id}": {
      "get": {
        "tags": [
          "Objects"
        ],
        "summary": "Get object with ID",
        "description": "Get a specific object with ID in MongoDB",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Found object with ID: ID"
                    },
                    "data": {
                      "type": "array",
                      "example": [
                        {
                          "_id": "61e4f519d450770ffab60d9c",
                          "objectName": "Object",
                          "objectURL": "https://sample-url.com",
                          "scale": "50 50 50",
                          "rotation": "0 0 0",
                          "position": "50 0 50",
                          "__v": 0
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error message"
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/environments/": {
      "get": {
        "tags": [
          "Environments"
        ],
        "summary": "Get a list of environments",
        "description": "Get a list of environments in MongoDB",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Retreived all environments"
                    },
                    "count": {
                      "type": "integer",
                      "example": 1
                    },
                    "data": {
                      "type": "array",
                      "example": [
                        {
                          "_id": "61e4f519d450770ffab60d9c",
                          "environmentName": "Object",
                          "environmentURL": "https://sample-url.com",
                          "__v": 0
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error message"
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/environments/{id}": {
      "get": {
        "tags": [
          "Environments"
        ],
        "summary": "Get environment with ID",
        "description": "Get a specific environment with ID in MongoDB",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Found environment with ID: ID"
                    },
                    "data": {
                      "type": "array",
                      "example": [
                        {
                          "_id": "61e4f519d450770ffab60d9c",
                          "objectName": "Environment",
                          "objectURL": "https://sample-url.com",
                          "__v": 0
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error message"
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}